[11,35,irods/irods_8102,irods/irods_8330] Update for new CMake, bump fuse
Newer versions of CMake have removed compatibility with older versions of
CMake. Given that this project specifies a minimum required CMake of 3.5, some
updates are needed.

This required a complete rewrite of the find module for fuse, as it was
written for CMake 2, which resulted in the addition of a find module for
librt.